feat: create `wrapMultiPlugin` to allow wrapping plugins like `publish`
Presently, some `semantic-release` plugins can be defined as arrays of
plugins. Wrapping requires a special approach because
`semantic-release` handles that format differently internally, allowing
for the plugins to be using a reducing function while running in
sequence.

Trying to mimic the reducing function behavior introduces a lot of
complexity and makes it difficult to guarantee the same behavior unless
the code is lifted, which in turn is a direct dependency on the
internals of `semantic-release`. No thanks.

To work around this, we hook into the special treatment of arrays by
defining the plugin as an array of decorated plugins. Since we don’t
know beforehand how many plugins we’re expected to wrap, we use a size
10 array as a “catch all”. Any “unused” plugin wrappers are effectively
a no-op.

The catch, as with `wrapPlugin`, is that any plugin-specific options
have to be defined inside a namespace in the release config.